10-socket_io/requirements.txt
python-socketio
websocket-client
eventlet
requests

pygame